+++ src/expand/mod.rs
@@ -150,23 +150,43 @@ where
     Parsing(parse::Error<Input>),
 
     /// Expansion error.
-    #[display(fmt = "Regex expansion failed: {}", _0)]
-    Expansion(UnknownParameterError<Input>),
+    #[display(fmt = "Failed to expand regex: {}", _0)]
+    Expansion(ParameterError<Input>),
 
     /// [`Regex`] creation error.
     #[display(fmt = "Regex creation failed: {}", _0)]
     Regex(regex::Error),
 }
 
-/// Error of an unknown [`Parameter`] being used in an [`Expression`].
-#[derive(Clone, Copy, Debug, Display, Error)]
-#[display(fmt = "Parameter '{}' not found.", not_found)]
-pub struct UnknownParameterError<Input>
+/// Possible [`Parameter`] errors being used in an [`Expression`].
+#[derive(Clone, Debug, Display, Error)]
+pub enum ParameterError<Input>
 where
     Input: fmt::Display,
 {
     /// [`Parameter`] not found.
-    pub not_found: Input,
+    #[display(fmt = "Parameter '{}' not found.", _0)]
+    NotFound(Input),
+
+    /// Failed to rename [`Regex`] capturing group.
+    #[display(
+        fmt = "Failed to rename groups in regex '{}' of parameter '{}': {}",
+        re,
+        parameter,
+        err
+    )]
+    RenameRegexGroup {
+        /// [`Parameter`] name.
+        parameter: Input,
+
+        /// [`Regex`] of the [`Parameter`].
+        re: String,
+
+        /// [`Error`] of parsing [`Regex`].
+        ///
+        /// [`Error`]: regex_syntax::Error
+        err: regex_syntax::Error,
+    },
 }
 
 /// Expansion of a [Cucumber Expressions][0] [AST] element into a [`Regex`] by
@@ -177,7 +197,7 @@ where
 /// [AST]: https://en.wikipedia.org/wiki/Abstract_syntax_tree
 pub trait IntoRegexCharIter<Input: fmt::Display> {
     /// Type of an [`Iterator`] performing the expansion.
-    type Iter: Iterator<Item = Result<char, UnknownParameterError<Input>>>;
+    type Iter: Iterator<Item = Result<char, ParameterError<Input>>>;
 
     /// Consumes this [AST] element returning an [`Iterator`] over [`char`]s
     /// transformable into a [`Regex`].
@@ -208,7 +228,7 @@ where
 /// [`IntoRegexCharIter::Iter`] for an [`Expression`].
 type ExpressionIter<Input> = iter::Chain<
     iter::Chain<
-        iter::Once<Result<char, UnknownParameterError<Input>>>,
+        iter::Once<Result<char, ParameterError<Input>>>,
         iter::FlatMap<
             vec::IntoIter<SingleExpression<Input>>,
             <SingleExpression<Input> as IntoRegexCharIter<Input>>::Iter,
@@ -218,7 +238,7 @@ type ExpressionIter<Input> = iter::Chain<
                 -> <SingleExpression<Input> as IntoRegexCharIter<Input>>::Iter,
         >,
     >,
-    iter::Once<Result<char, UnknownParameterError<Input>>>,
+    iter::Once<Result<char, ParameterError<Input>>>,
 >;
 
 impl<Input> IntoRegexCharIter<Input> for SingleExpression<Input>
@@ -307,7 +327,7 @@ type AlternationIter<I> = iter::Chain<
             >,
         >,
     >,
-    iter::Once<Result<char, UnknownParameterError<I>>>,
+    iter::Once<Result<char, ParameterError<I>>>,
 >;
 
 // TODO: Replace with TAIT, once stabilized:
@@ -319,7 +339,7 @@ type AlternationIterInner<I> = iter::Chain<
         <Alternative<I> as IntoRegexCharIter<I>>::Iter,
         fn(Alternative<I>) -> <Alternative<I> as IntoRegexCharIter<I>>::Iter,
     >,
-    iter::Once<Result<char, UnknownParameterError<I>>>,
+    iter::Once<Result<char, ParameterError<I>>>,
 >;
 
 impl<Input> IntoRegexCharIter<Input> for Alternative<Input>
@@ -397,7 +417,7 @@ type OptionalIter<Input> = iter::Map<
 >;
 
 /// Function pointer describing [`Ok`].
-type MapOkChar<Input> = fn(char) -> Result<char, UnknownParameterError<Input>>;
+type MapOkChar<Input> = fn(char) -> Result<char, ParameterError<Input>>;
 
 impl<Input> IntoRegexCharIter<Input> for Parameter<Input>
 where
@@ -434,25 +454,20 @@ where
         } else if eq(&self.input, "word") {
             Left(Left(r#"([^\s]+)"#.chars().map(Ok)))
         } else if eq(&self.input, "string") {
-            // We add `string_<id>` name to `string` parameters to later parse
-            // out leading and trailing quotes in `cucumber-codegen` crate.
-            // See https://github.com/cucumber-rs/cucumber-expressions/issues/7
-            // for more info.
             Left(Right(
                 OwnedChars::new(format!(
-                    "(?P<string_{}>\
-                      \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"\
-                      |'(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)')",
-                    self.id,
+                    "(?:\
+                      \"(?P<__{id}_0>[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"\
+                      |'(?P<__{id}_1>[^'\\\\]*(?:\\\\.[^'\\\\]*)*)'\
+                    )",
+                    id = self.id,
                 ))
                 .map(Ok),
             ))
         } else if eq(&self.input, "") {
             Left(Left(r#"(.*)"#.chars().map(Ok)))
         } else {
-            Right(iter::once(Err(UnknownParameterError {
-                not_found: self.input,
-            })))
+            Right(iter::once(Err(ParameterError::NotFound(self.input))))
         }
     }
 }
@@ -464,14 +479,11 @@ type ParameterIter<Input> = Either<
     Either<
         iter::Map<
             str::Chars<'static>,
-            fn(char) -> Result<char, UnknownParameterError<Input>>,
+            fn(char) -> Result<char, ParameterError<Input>>,
         >,
-        iter::Map<
-            OwnedChars,
-            fn(char) -> Result<char, UnknownParameterError<Input>>,
-        >,
+        iter::Map<OwnedChars, fn(char) -> Result<char, ParameterError<Input>>>,
     >,
-    iter::Once<Result<char, UnknownParameterError<Input>>>,
+    iter::Once<Result<char, ParameterError<Input>>>,
 >;
 
 /// [`Iterator`] for skipping a last [`Item`].
@@ -630,7 +642,7 @@ where
 // Naming of test cases is preserved.
 #[cfg(test)]
 mod spec {
-    use super::{Error, Expression, UnknownParameterError};
+    use super::{Error, Expression, ParameterError};
 
     #[test]
     fn alternation_with_optional() {
@@ -743,9 +755,10 @@ mod spec {
 
         assert_eq!(
             expr.as_str(),
-            "^(?P<string_0>\
-               \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
-               '(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)')$",
+            "^(?:\
+                \"(?P<__0_0>[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
+                '(?P<__0_1>[^'\\\\]*(?:\\\\.[^'\\\\]*)*)'\
+             )$",
         );
         assert!(expr.is_match("\"\""));
         assert!(expr.is_match("''"));
@@ -764,12 +777,13 @@ mod spec {
 
         assert_eq!(
             expr.as_str(),
-            "^(?P<string_0>\
-               \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
-               '(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)') \
-             (?P<string_1>\
-               \"(?:[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
-               '(?:[^'\\\\]*(?:\\\\.[^'\\\\]*)*)')$",
+            "^(?:\
+                \"(?P<__0_0>[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
+                '(?P<__0_1>[^'\\\\]*(?:\\\\.[^'\\\\]*)*)'\
+              ) (?:\
+                \"(?P<__1_0>[^\"\\\\]*(?:\\\\.[^\"\\\\]*)*)\"|\
+                '(?P<__1_1>[^'\\\\]*(?:\\\\.[^'\\\\]*)*)'\
+              )$",
         );
         assert!(expr.is_match("\"\" ''"));
         assert!(expr.is_match("'' \"\""));
@@ -816,10 +830,10 @@ mod spec {
     #[test]
     fn unknown_parameter() {
         match Expression::regex("{custom}").unwrap_err() {
-            Error::Expansion(UnknownParameterError { not_found }) => {
+            Error::Expansion(ParameterError::NotFound(not_found)) => {
                 assert_eq!(*not_found, "custom");
             }
-            e @ (Error::Parsing(_) | Error::Regex(_)) => {
+            e @ (Error::Parsing(_) | Error::Regex(_) | Error::Expansion(_)) => {
                 // TODO: Use "{e}" syntax once MSRV bumps above 1.58.
                 panic!("wrong err: {}", e);
             }